Refactoring some of the code here
